+++ packages/rn/theme/themes/index.ts
@@ -1,12 +1,19 @@
 import { dark } from "./dark";
 import { light } from "./light";
-import { grapeLight, grapeDark } from "./grape";
-import { peachLight, peachDark } from "./peach";
-import { retroLight, retroDark } from "./retro";
-import { monoLight, monoDark } from "./mono";
-import { lavenderLight, lavenderDark } from "./lavender";
-import { oceanLight, oceanDark } from "./ocean";
-import { bubblegumLight, bubblegumDark } from "./bubblegum";
+import { grapeLight } from "./grape-light";
+import { grapeDark } from "./grape-dark";
+import { peachLight } from "./peach-light";
+import { peachDark } from "./peach-dark";
+import { retroLight } from "./retro-light";
+import { retroDark } from "./retro-dark";
+import { monoLight } from "./mono-light";
+import { monoDark } from "./mono-dark";
+import { lavenderLight } from "./lavender-light";
+import { lavenderDark } from "./lavender-dark";
+import { oceanLight } from "./ocean-light";
+import { oceanDark } from "./ocean-dark";
+import { bubblegumLight } from "./bubblegum-light";
+import { bubblegumDark } from "./bubblegum-dark";
 
 export const themes = {
   light,
